refactor: 상품 리스트 페이지 컴포넌트로 분리
